+++ Contents/mods/ChaosMod/42/media/lua/client/ChaosMod/Effects/List/EffectSpawnRandomAnimal.lua
@@ -20,11 +20,8 @@ function EffectSpawnRandomAnimal:OnStart()
         local breedName = getText("IGUI_Breed_" .. animal:getData():getBreed():getName())
 
         local str = string.format(ChaosLocalization.GetString("misc", "spawned_animal"), breedName, displayName)
-        ChaosPlayer.SayLine(player, str, 0.05, 0.333, 1.0)
+        ChaosPlayer.SayLine(player, str, 162 / 255, 223 / 255, 208 / 255)
 
-        table.insert(ChaosMod.specialAnimalsFollowers, {
-            animal = animal,
-            repathTicks = 20
-        })
+        SpecialAnimal:new(animal)
     end
 end